requirements.txt
streamlit
langchain.agent
langchain.chat_models
langchain.chains.conversation.memory
transformers
langchain.tools